Backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-cors==4.0.1
anthropic==0.28.0
pdfplumber==0.11.2
python-docx==1.1.2
#psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
werkzeug==3.0.3
gunicorn==22.0.0